+++ packages/ProgressIndicators/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.4](/compare/@onaio/progress-indicators@0.0.3...@onaio/progress-indicators@0.0.4) (2020-08-20)
+
+**Note:** Version bump only for package @onaio/progress-indicators
+
 ## [0.0.3](/compare/@onaio/progress-indicators@0.0.2...@onaio/progress-indicators@0.0.3) (2020-03-04)
 
 **Note:** Version bump only for package @onaio/progress-indicators